resurrect: compile+link+run on modern nightly, portable across glibc/musl and x86_64/aarch64
Modern-nightly (removed features):
- llvm_asm! -> compiler_fence / .swap_bytes() / ptr::write_bytes
- VaListImpl->VaList, .arg->.next_arg, drop .as_va_list; drop stabilized feature gates
- wrapping_offset_from->offset_from; packed field refs -> &raw; bitfield methods via aligned copy
- stime -> clock_settime shim; __res_mkquery -> res_mkquery

Portability (arch + libc):
- st_nlink / tv_sec width casts; ioctl request width via 'as _' (Ioctl differs glibc/musl)
- termios c_ispeed/c_ospeed -> cfg-gated __c_ispeed on musl
- c_char signedness: data-table bytes as 'Nu8 as libc::c_char' (u8 aarch64 / i8 x86_64)
- musl-only #[no_mangle] shims: gnu_dev_major/minor/makedev, __cmsg_nxthdr

+++ archival/cpio.rs
@@ -182,7 +182,7 @@ unsafe fn cpio_o() -> libc::c_int {
         }
         /* Store hardlinks for later processing, dont output them */
         if !(st.st_mode & 0o170000i32 as libc::c_uint == 0o40000i32 as libc::c_uint)
-          && st.st_nlink > 1i32 as libc::c_ulong
+          && st.st_nlink as libc::c_ulong > 1i32 as libc::c_ulong
         {
           let mut n: *mut name_s = std::ptr::null_mut();
           let mut l: *mut inodes_s = std::ptr::null_mut();
